+++ packages/browser-extension/src/store/index.ts
@@ -1,1 +1,2 @@
 export * from "./store";
+export * from "./hooks";
