+++ packages/web-app/package.json
@@ -21,8 +21,8 @@
     "type-check": "npm run prisma:generate && tsc --noEmit"
   },
   "dependencies": {
+    "@aws-sdk/client-ssm": "^3.693.0",
     "@prisma/client": "^5.0.0",
-    "@react-oauth/google": "^0.12.1",
     "@reduxjs/toolkit": "^2.3.0",
     "deep-equal": "^2.2.3",
     "express": "^4.18.2",
