+++ packages/web-app/src/app/page.tsx
@@ -4,6 +4,7 @@ import React, { useEffect } from "react";
 import { View } from "react-native";
 import { Text, Surface, List, ActivityIndicator } from "react-native-paper";
 import { Provider } from "react-redux";
+
 import { store } from "../store/store";
 import { useAppDispatch, useAppSelector } from "../store/hooks";
 import { fetchArgumentMaps } from "../store/argumentMapsSlice";
